need to make a popup or some kind of interrupt for when there are no more spaces
