+++ chip.css
@@ -86,6 +86,7 @@ html {
   margin: 10px;
 }
 
+
 /* #start_btn{
     border-radius: 10px;
     background-color: green;
